+++ CHANGELOG.md
@@ -39,3 +39,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 1.3.3
 
 - fix: default options errors
+- fix: revert core-js
+- fix: revert sass version
+- fix: add `col` field options to docs